+++ spigot/src/main/kotlin/com/dreamdisplays/commands/DisplayCommand.kt
@@ -26,6 +26,7 @@ class DisplayCommand :
         CreateCommand(),
         DeleteCommand(),
         VideoCommand(),
+        InfoCommand(),
         ListCommand(),
         StatsCommand(),
         ReloadCommand(),